trust: separate crate into modules
