+++ crates/trust/src/lib.rs
@@ -1,4 +1,7 @@
+#[path = "trust/trust.rs"]
 pub mod trust;
+
+#[path = "trust_db/trust_db.rs"]
 pub mod trust_db;
 
 pub use trust::{Trust, TrustConfig, TrustError};
